jwt fix

+++ src/templateRoutes.py
@@ -2,9 +2,9 @@ from flask import Blueprint, request
 from flask.json import jsonify
 from src.db import db1
 from bson import ObjectId
-from flask_jwt_extended import get_jwt_identity, jwt_required
+from flask_jwt_extended import  get_jwt_identity, jwt_required
 from flasgger import swag_from
-from src.constants.http_status_codes import HTTP_200_OK, HTTP_201_CREATED, HTTP_404_NOT_FOUND
+from src.constants.http_status_codes import HTTP_200_OK, HTTP_201_CREATED, HTTP_404_NOT_FOUND, HTTP_401_UNAUTHORIZED
 
 
 templateRoutes = Blueprint("templateRoutes", __name__, url_prefix="/template")
@@ -13,12 +13,13 @@ templateRoutes = Blueprint("templateRoutes", __name__, url_prefix="/template")
 @jwt_required()
 @swag_from('./docs/template/createTemplate.yaml')
 def create_template():
+    current_user = get_jwt_identity()
 
     template_name = request.json["template_name"]
     subject = request.json["subject"]
     body = request.json["body"]
    
-    res = db1.insert_one({"template_name":template_name,"subject":subject,"body":body})
+    res = db1.insert_one({"template_name":template_name,"subject":subject,"body":body, "creator": current_user})
     return jsonify({
     'message': "Template created",
     
@@ -30,10 +31,13 @@ def create_template():
 @jwt_required()
 @swag_from('./docs/template/getAllTemplate.yaml')
 def get_templates():
+    current_user = get_jwt_identity()
+    
     if request.method == "GET":
         templates =[]
         for i in db1.find():
-            templates.append({"_ID":str(ObjectId(i["_id"])),"template_name":i["template_name"],"subject":i["subject"],"body":i["body"]})
+            if i["creator"] == current_user:
+                templates.append({"_ID":str(ObjectId(i["_id"])),"template_name":i["template_name"],"subject":i["subject"],"body":i["body"]})
         return jsonify(templates), HTTP_200_OK
    
 
@@ -42,17 +46,22 @@ def get_templates():
 @jwt_required()
 @swag_from('./docs/template/updateTemplate.yaml')
 def update_template(id):
+    current_user = get_jwt_identity()
    
     template = db1.find_one({"_id":ObjectId(id)})
+    print(template["subject"])
     if not template:
         return jsonify({'message': 'Item not found'}), HTTP_404_NOT_FOUND
 
-    db1.update_one({"_id":ObjectId(id)},{"$set":{
-        "template_name":request.json["template_name"],
-        "subject":request.json["subject"],
-        "body":request.json["body"]
-    }})
-    return jsonify({"message":"updated!"}), HTTP_200_OK
+    if template["creator"] == current_user:
+
+        db1.update_one({"_id":ObjectId(id)},{"$set":{
+            "template_name":request.json["template_name"],
+            "subject":request.json["subject"],
+            "body":request.json["body"]
+        }})
+        return jsonify({"message":"updated!"}), HTTP_200_OK
+    return jsonify({'message': 'No access!'}), HTTP_401_UNAUTHORIZED
 
 
 
@@ -60,14 +69,17 @@ def update_template(id):
 @jwt_required()
 @swag_from('./docs/template/deleteTemplate.yaml')
 def delete_template(id):
+    current_user = get_jwt_identity()
 
     template = db1.find_one({"_id":ObjectId(id)})
     if not template:
         return jsonify({'message': 'Item not found'}), HTTP_404_NOT_FOUND
     
+    if template["creator"] == current_user:
     
-    db1.delete_one({"_id":ObjectId(id)})
-    return jsonify({"message":"deleted!"}), HTTP_200_OK
+        db1.delete_one({"_id":ObjectId(id)})
+        return jsonify({"message":"deleted!"}), HTTP_200_OK
+    return jsonify({'message': 'No access!'}), HTTP_401_UNAUTHORIZED
 
     
    
@@ -77,19 +89,23 @@ def delete_template(id):
 @jwt_required()
 @swag_from('./docs/template/getTemplate.yaml')
 def get_template(id):
+    current_user = get_jwt_identity()
+    
     template = db1.find_one({"_id":ObjectId(id)})
 
     if not template:
         return jsonify({'message': 'Item not found'}), HTTP_404_NOT_FOUND
     
-    return jsonify({
-                        'message': 'Success',
-                        'template': {
-                            "_ID":str(ObjectId(template["_id"])),
-                            "template_name":template["template_name"],
-                            "subject":template["subject"],
-                            "body":template["body"]
-                        }
+    if template["creator"] == current_user:
+        return jsonify({
+                            'message': 'Success',
+                            'template': {
+                                "_ID":str(ObjectId(template["_id"])),
+                                "template_name":template["template_name"],
+                                "subject":template["subject"],
+                                "body":template["body"]
+                            }
 
-                    }), HTTP_200_OK
+                        }), HTTP_200_OK
+    return jsonify({'message': 'No access!'}), HTTP_401_UNAUTHORIZED
